+++ src/BaseBinaryStar.cpp
@@ -1738,9 +1738,9 @@ void BaseBinaryStar::ResolveMainSequenceMerger() {
     double initialHydrogenFraction = m_Star1->InitialHydrogenAbundance();
     
     double finalHydrogenMass;
-    if ((OPTIONS->MainSequenceCoreMassPrescription() == CORE_MASS_PRESCRIPTION::SHIKAUCHI) &&
-        (utils::Compare(m_Star1->MZAMS(), SHIKAUCHI_LOWER_MASS_LIMIT) >= 0)                &&
-        (utils::Compare(m_Star2->MZAMS(), SHIKAUCHI_LOWER_MASS_LIMIT) >= 0)) {
+    if ((OPTIONS->MainSequenceCoreMassPrescription() == CORE_MASS_PRESCRIPTION::BRCEK) &&
+        (utils::Compare(m_Star1->MZAMS(), BRCEK_LOWER_MASS_LIMIT) >= 0)                &&
+        (utils::Compare(m_Star2->MZAMS(), BRCEK_LOWER_MASS_LIMIT) >= 0)) {
         
         double coreMass1       = m_Star1->MainSequenceCoreMass();
         double coreMass2       = m_Star2->MainSequenceCoreMass();
@@ -2118,8 +2118,8 @@ void BaseBinaryStar::CalculateMassTransfer(const double p_Dt) {
                                                                                  donorIsHeRich);
     double massDiffDonor            = 0.0;
         
-    // can the mass transfer happen on a nuclear timescale?  only considering this for MS donors
-    if (m_Donor->IsOneOf(ALL_MAIN_SEQUENCE)) {
+    // can the mass transfer happen on a nuclear timescale?
+    if (m_Donor->IsOneOf(NON_COMPACT_OBJECTS)) {
         // technically, we do not know how much mass the accretor should gain until we do the calculation, which impacts the RL size, so we will check whether a nuclear timescale MT was feasible later
         double maximumAccretedMass = maximumAccretionRate * m_Dt;
         if (OPTIONS->MassTransferAccretionEfficiencyPrescription() == MT_ACCRETION_EFFICIENCY_PRESCRIPTION::THERMALLY_LIMITED) {
@@ -2132,7 +2132,7 @@ void BaseBinaryStar::CalculateMassTransfer(const double p_Dt) {
         }
         
         // check that the star really would have consistently fit into the Roche lobe
-        double zetaEquilibrium = m_Donor->CalculateZetaEquilibrium();                                                               // note: value is only meaningful for MS donor
+        double zetaEquilibrium = m_Donor->CalculateZetaEquilibrium();
         double zetaLobe = CalculateZetaRocheLobe(jLoss, m_FractionAccreted);
         if (utils::Compare(zetaEquilibrium, zetaLobe) > 0  && massDiffDonor > 0.0) {                                                // yes, it's nuclear timescale mass transfer; no need for utils::Compare here
             m_MassLossRateInRLOF    = massDiffDonor / m_Dt;